Move ListRangeIterator to utils package
